+++ crates/wdl-engine/CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+## 0.13.0 - 2026-03-01
+
 #### Changed
 
 * The Slurm (with Apptainer) backend now uses `sbatch` to queue new jobs